Prefer hashes over objects when storing in session
Currently the gem stores whole objects in the session storage. This works as
long as the gem consumer app uses the `marshal` cookie serializer.

Since rails 4.1 the default and preferred serializer is `json`.

For the gem to be compatible with all rails versions and all serializers,
prefer storing hashes instead and objects to the session. This makes it easier
and cleaner to store as well as load from session.