Replace Digest::HMAC with OpenSSL::HMAC
Digest::HMAC was part of ruby's standard library up to version 2.1.
Even then, it had a giant warning:

https://github.com/ruby/ruby/blob/v2_1_0/ext/digest/lib/digest/hmac.rb

"CAUTION: Use of this library is discouraged, because this
implementation was meant to be experimental but somehow got into the 1.9
series without being noticed. Please use OpenSSL::HMAC in the “openssl”
library instead."

This gem still managed to find Digest::HMAC though, because oauth-ruby
(yogurt dependency) "generously" provides an old version taken from the
ruby standard library as it can be seen here:
https://github.com/oauth-xx/oauth-ruby/blob/v0.4.7/lib/digest/hmac.rb

This commit makes use of the openssl implementation of HMAC and it
should work both in ruby 2.3 & ruby 2.5.